+++ apps/mobile/ios/Keys/KeysCapture.m
@@ -13,4 +13,8 @@ RCT_EXTERN_METHOD(capture
                   : (RCTPromiseResolveBlock)resolve rejecter
                   : (RCTPromiseRejectBlock)reject)
 
+RCT_EXTERN_METHOD(whereAmI
+                  : (RCTPromiseResolveBlock)resolve rejecter
+                  : (RCTPromiseRejectBlock)reject)
+
 @end